backend/requirements.txt

fastapi
uvicorn
crewai
crewai-tools
langchain-groq
sqlalchemy
pydantic
python-dotenv
PyGithub
python-multipart
httpx